Add ft_fnmatch tests
There are some diffences between ft_fnmatch and fnmatch, in edge cases
handling. Especially for invalid ranges. If a range is invalid like z-a
for example, it is simply skipped, whereas, due to implementation,
ft_fnmatch tests the first character of the range which can result in a
match. This wouldn't happen in the original (which means the range is
parsed before the first character is tested, not the other way around).
Also, if the range is empty on one side or more, it just performs
matching as they were regular characters. In ft_fnmatch, this wouldn't
happen because the '-' character is always special inside brackets and
needs to be escaped to be tested as a regular character.

These differences can be ignored for now, as the two functions only
differ on invalid globbing syntax which results in undefined behavior.