Fix two or three things so that globings work
I mean, this is probably still very wrong on multiple levels, but it
basically works for now. The next step is to crash test it, makes sure
it behaves like the original in a lot of different situations, not
matter how tricky (except maybe on things like NO_SORT, and other non
shell-essential functions). Test that it doesnt leak too much also.

Then, and only then, the brackets can be implemented.